OF-2383: Move tests of the AbstractGroupProvider to its own test class, and expand
