+++ xmppserver/src/main/java/org/jivesoftware/openfire/group/AbstractGroupProvider.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright (C) 2022 Ignite Realtime Foundation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.jivesoftware.openfire.group;
 
 import java.io.Serializable;
@@ -10,16 +25,8 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.locks.Lock;
 
-import com.google.common.collect.Interner;
-import com.google.common.collect.Interners;
-import org.apache.commons.lang3.StringUtils;
 import org.jivesoftware.database.DbConnectionManager;
-import org.jivesoftware.openfire.XMPPServer;
-import org.jivesoftware.openfire.event.GroupEventDispatcher;
-import org.jivesoftware.openfire.event.GroupEventListener;
-import org.jivesoftware.util.CacheableOptional;
 import org.jivesoftware.util.SystemProperty;
 import org.jivesoftware.util.cache.Cache;
 import org.jivesoftware.util.cache.CacheFactory;
@@ -274,7 +281,7 @@ public abstract class AbstractGroupProvider implements GroupProvider {
                 if (DbConnectionManager.getDatabaseType() == DbConnectionManager.DatabaseType.oracle) {
                     pstmt = con.prepareStatement("SELECT 1 FROM dual " + HAS_SHARED_GROUPS_PARTIAL);
                 } else {
-                    pstmt = con.prepareStatement("SELECT 1 " + HAS_SHARED_GROUPS_PARTIAL);
+                    pstmt = con.prepareStatement("SELECT 1 FROM ofGroupProp " + HAS_SHARED_GROUPS_PARTIAL);
                 }
                 rs = pstmt.executeQuery();
                 result = rs.next();
