+++ xmppserver/src/test/java/org/jivesoftware/openfire/group/DefaultGroupProviderTest.java
@@ -405,1204 +405,6 @@ public class DefaultGroupProviderTest extends DBTestCase
     }
 
     /**
-     * Verifies that a {@link DefaultGroupProvider#getPublicSharedGroupNames()} returns nothing when there
-     * are no groups that are shared with everyone.
-     */
-    public void testPublicSharedGroupNamesNone() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getPublicSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getPublicSharedGroupNames()} returns one group when there
-     * is one groups that is shared with everyone.
-     */
-    public void testPublicSharedGroupNamesOne() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithEverybody("Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getPublicSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getPublicSharedGroupNames()} returns all group names when
-     * there are multiple groups that are shared with everyone.
-     */
-    public void testPublicSharedGroupNamesMultiple() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithEverybody("Users in group A");
-        provider.createGroup("Test Group B").shareWithEverybody("Users in group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getPublicSharedGroupNames();
-
-        // Verify result.
-        assertEquals(2, result.size());
-        assertTrue(result.contains("Test Group A"));
-        assertTrue(result.contains("Test Group B"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getPublicSharedGroupNames()} returns nothing when there
-     * are no groups that are shared with everyone, but when there is a group that is shared explicitly with 'nobody'.
-     */
-    public void testPublicSharedGroupNamesWithNobody() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithNobody();
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getPublicSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getPublicSharedGroupNames()} returns nothing when there
-     * are no groups that are shared with everyone, but when there is a group that is shared explicitly with members
-     * of that group.
-     */
-    public void testPublicSharedGroupNamesWithOnlyGroup() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getPublicSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} returns nothing when there are no groups
-     * that are shared.
-     */
-    public void testSharedGroupNamesNone() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} returns nothing when there are no groups
-     * that are shared, but one that's explicitly shared with 'nobody'.
-     */
-    public void testSharedGroupNamesNobody() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithNobody();
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} returns nothing a group that is shared with
-     * everyone.
-     */
-    public void testSharedGroupNamesOneWithEveryone() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithEverybody("Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} returns nothing a group that is shared with
-     * members of that group.
-     */
-    public void testSharedGroupNamesOneWithGroup() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} returns nothing a group that is shared with
-     * everyone and another group that is shared explicitly with members of that group.
-     */
-    public void testSharedGroupNamesOneWithMix() throws Exception
-    {
-        // Setup test fixture.
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-        provider.createGroup("Test Group B").shareWithEverybody("Users in group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(2, result.size());
-        assertTrue(result.contains("Test Group A"));
-        assertTrue(result.contains("Test Group B"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationEverybodyToEverybody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Users in group A"); // Technically not a 'change' - but should not generate false results anyway!
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationEverybodyToEverybodyRename() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Users in group A RENAMED"); // Different name
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationEverybodyToSameGroup() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationEverybodyToOtherGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to no longer be shared.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationEverybodytoSharedWithNobody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithNobody();
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSameGroupToEverybody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Test Group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSameGroupToSameGroup() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A"); // Technically not a 'change' - but should not generate false results anyway!
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSameGroupRenamed() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A RENAMED"); // Different name
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSameGroupToOtherGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to no longer be shared.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSameGrouptoSharedWithNobody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithNobody();
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherGroupToEverybody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Test Group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherGroupToSameGroup() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherRename() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A RENAMED"); // Different name
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherToOtherGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A"); // Technically not a 'change' - but should not generate false results anyway!
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherToAdditionalGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-        provider.createGroup("Test Group C");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Arrays.asList("Test Group B", "Test Group C"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherToFewerGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Arrays.asList("Test Group B", "Test Group C"), "Users in group A");
-        provider.createGroup("Test Group B");
-        provider.createGroup("Test Group C");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to no longer be shared.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationOtherGrouptoSharedWithNobody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithNobody();
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has been removed.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterGroupRemoval() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithEverybody("Users in group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        provider.deleteGroup(group.getName());
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationUnsharedToEverybody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Test Group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationUnsharedToSameGroup() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationUnsharedToOtherGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to no longer be shared.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationUnsharedSharedWithNobody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithNobody();
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSharedWithNobodyToEverybody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithNobody();
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithEverybody("Test Group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSharedWithNobodyToSameGroup() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithNobody();
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInSameGroup("Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to be shared with a different set of users.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSharedWithNobodyToOtherGroups() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithNobody();
-        provider.createGroup("Test Group B");
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames()} does not return a stale cache entry after a
-     * previously obtained group has reconfigured to no longer be shared.
-     */
-    public void testSharedGroupNamesStaleCacheTestAfterModificationSharedWithNobodytoSharedWithNobody() throws Exception
-    {
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithNobody();
-
-        // Execute system under test;
-        provider.getSharedGroupNames(); // populate cache.
-        group.shareWithNobody(); // Technically not a 'change' - but should not generate false results anyway!
-        final Collection<String> result = provider.getSharedGroupNames();
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, without the user being in any of the groups.
-     */
-    public void testSharedGroupNamesByNameNone() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, with the user being in a member of one of the groups.
-     */
-    public void testSharedGroupNamesByNameNoneButMember() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B").getMembers().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, with the user being in an admin of one of the groups.
-     */
-    public void testSharedGroupNamesByNameNoneButAdmin() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        provider.createGroup("Test Group B").getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, but one that is explicitly shared with 'nobody', without the user being in any of the groups.
-     */
-    public void testSharedGroupNamesByNameNobody() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithNobody();
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, but one that is explicitly shared with 'nobody', with the user being in a member of that group.
-     */
-    public void testSharedGroupNamesByNameNobodyButMember() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        final Group input = provider.createGroup("Test Group B");
-        input.shareWithNobody();
-        input.getMembers().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there are no groups
-     * that are shared, but one that is explicitly shared with 'nobody', with the user being in an admin of that group.
-     */
-    public void testSharedGroupNamesByNameNobodyButAdmin() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        final Group input = provider.createGroup("Test Group B");
-        input.shareWithNobody();
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there are is one
-     * group that is shared with everybody, without the user being in that group.
-     */
-    public void testSharedGroupNamesByNameWithEveryoneNoAssociation() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithEverybody("Users in group A");
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there are is one
-     * group that is shared with everybody, with the user being in a member of that group.
-     */
-    public void testSharedGroupNamesByNameWithEveryoneMember() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        final Group input = provider.createGroup("Test Group B");
-        input.shareWithEverybody("Users in group B");
-        input.getMembers().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group B"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there are is one
-     * group that is shared with everybody, with the user being in an admin of that group.
-     */
-    public void testSharedGroupNamesByNameWithEveryoneAdmin() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A");
-        final Group input = provider.createGroup("Test Group B");
-        input.shareWithEverybody("Users in group B");
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group B"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns nothing when there's a group
-     * shared with users in the group, without the user being in that group.
-     */
-    public void testSharedGroupNamesByNameWithGroupNoAssociation() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there are is one
-     * group shared with users in the group, with the user being in a member of that group.
-     */
-    public void testSharedGroupNamesByNameWithGroupMember() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group input = provider.createGroup("Test Group A");
-        input.shareWithUsersInSameGroup("Users in group A");
-        input.getMembers().add(needle);
-
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there are is one
-     * group shared with users in the group, with the user being in an admin of that group.
-     */
-    public void testSharedGroupNamesByNameWithGroupAdmin() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group input = provider.createGroup("Test Group A");
-        input.shareWithUsersInSameGroup("Users in group A");
-        input.getAdmins().add(needle);
-
-        provider.createGroup("Test Group B");
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns no group when there are is one
-     * group that is shared with users in the group, with the user being in a member of another group.
-     */
-    public void testSharedGroupNamesByNameWithGroupMemberOfOtherGroup() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-
-        final Group input = provider.createGroup("Test Group B");
-        input.getMembers().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns no group when there are is one
-     * group that is shared with users in the group, with the user being in an admin of another group.
-     */
-    public void testSharedGroupNamesByNameWithGroupAdminOfOtherGroup() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A").shareWithUsersInSameGroup("Users in group A");
-
-        final Group input = provider.createGroup("Test Group B");
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns a group when there is one
-     * group that is shared with users in another group, with the user being in a member of that other group.
-     */
-    public void testSharedGroupNamesByNameWithOtherGroupMemberOfOtherGroup() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A") // Group A is shared with users in group B!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-
-        final Group input = provider.createGroup("Test Group B");
-        input.getMembers().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns no group when there are is one
-     * group that is shared with users in the group, with the user being in an admin of another group.
-     */
-    public void testSharedGroupNamesByNameWithOtherGroupAdminOfOtherGroup() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A") // Group A is shared with users in group B!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-
-        final Group input = provider.createGroup("Test Group B");
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns only groups that are configured to
-     * be directly visible to the (group of) the provided user, when recursion is disabled.
-     */
-    public void testSharedGroupNamesByNameRecursionDisabled() throws Exception
-    {
-        // Setup test fixture.
-        DefaultGroupProvider.SHARED_GROUP_RECURSIVE.setValue(false);
-
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A") // Group A is shared with users in group B!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-
-        provider.createGroup("Test Group B") // Group B is shared with users in group C!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group C"), "Users in group B");
-
-        provider.createGroup("Test Group C") // Group C is shared with users in group D!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group D"), "Users in group C");
-
-        final Group input = provider.createGroup("Test Group D");
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group C"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} returns groups that are configured to be
-     * directly visible to the (group of) the provided user, but also groups that are configured to be visible to
-     * members of those groups (etc), when recursion is enabled.
-     */
-    public void testSharedGroupNamesByNameRecursionEnabled() throws Exception
-    {
-        // Setup test fixture.
-        DefaultGroupProvider.SHARED_GROUP_RECURSIVE.setValue(true);
-
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        provider.createGroup("Test Group A") // Group A is shared with users in group B!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group B"), "Users in group A");
-
-        provider.createGroup("Test Group B") // Group B is shared with users in group C!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group C"), "Users in group B");
-
-        provider.createGroup("Test Group C") // Group C is shared with users in group D!
-            .shareWithUsersInGroups(Collections.singletonList("Test Group D"), "Users in group C");
-
-        final Group input = provider.createGroup("Test Group D");
-        input.getAdmins().add(needle);
-
-        // Execute system under test.
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(3, result.size());
-        assertTrue(result.contains("Test Group C"));
-        assertTrue(result.contains("Test Group B"));
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} does not return a stale cache test when
-     * a member is added to a shared group after the group as already loaded in a cache.
-     */
-    public void testSharedGroupNamesByNameStaleCacheTestMemberAdded() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in Group A");
-
-        // Execute system under test.
-        provider.getSharedGroupNames(needle); // populate cache
-        group.getMembers().add(needle);
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} does not return a stale cache test when
-     * a member is removed from a shared group after the group as already loaded in a cache.
-     */
-    public void testSharedGroupNamesByNameStaleCacheTestMemberRemoved() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.shareWithUsersInSameGroup("Users in Group A");
-        group.getMembers().add(needle);
-
-        // Execute system under test.
-        provider.getSharedGroupNames(needle); // populate cache
-        group.getMembers().remove(needle);
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} does not return a stale cache test when
-     * sharing is enabled after the group as already loaded in a cache.
-     */
-    public void testSharedGroupNamesByNameStaleCacheTestSharingEnabled() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.getMembers().add(needle);
-
-        // Execute system under test.
-        provider.getSharedGroupNames(needle); // populate cache
-        group.shareWithUsersInSameGroup("Users in Group A");
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(1, result.size());
-        assertTrue(result.contains("Test Group A"));
-    }
-
-    /**
-     * Verifies that a {@link DefaultGroupProvider#getSharedGroupNames(JID)} does not return a stale cache test when
-     * sharing is disabled after the group as already loaded in a cache.
-     */
-    public void testSharedGroupNamesByNameStaleCacheTestSharingDisabled() throws Exception
-    {
-        // Setup test fixture.
-        final JID needle = new JID("jane@example.org");
-        final DefaultGroupProvider provider = new DefaultGroupProvider();
-        final Group group = provider.createGroup("Test Group A");
-        group.getMembers().add(needle);
-        group.shareWithUsersInSameGroup("Users in Group A");
-
-        // Execute system under test.
-        provider.getSharedGroupNames(needle); // populate cache
-        group.shareWithNobody();
-        final Collection<String> result = provider.getSharedGroupNames(needle);
-
-        // Verify result.
-        assertEquals(0, result.size());
-    }
-
-    /**
      * Verifies that {@link DefaultGroupProvider#deleteGroup(String)} deletes the named group, such that it cannot be
      * retrieved
      */
